backend/requirements.txt
django-cors-headers == 3.13.0
Django == 4.1.4
psycopg2 == 2.9.6
djangorestframework == 3.13.1
python-dotenv
django-mptt == 0.14.0
djangorestframework-simplejwt == 5.2.2
Pillow == 9.0.0